+++ packages/common/src/PredicateFactory.ts
@@ -38,6 +38,8 @@ export class PredicateFactory {
 
   getRoot = (): string => this.root;
 
+  getGeneratedAt = (): number => this.generatedAt;
+
   getPredicateAddress = memoize((address: string | B256Address): string => {
     const predicateAddress = getFuelPredicateAddresses({
       signerAddress: this.adapter.convertAddress(address),
@@ -51,7 +53,7 @@ export class PredicateFactory {
       address: string | B256Address,
       provider: Provider,
       data?: T,
-    ): Predicate<T> =>
+    ): Predicate<InputValue[], { [name: string]: unknown }> =>
       new Predicate({
         bytecode: arrayify(this.bytecode),
         abi: this.abi,
@@ -60,7 +62,7 @@ export class PredicateFactory {
           SIGNER: this.adapter.convertAddress(address),
         },
         data,
-      }),
+      }) as Predicate<InputValue[], { [name: string]: unknown }>,
   );
 
   getAccountAddress = (