+++ packages/common/src/types.ts
@@ -36,7 +36,7 @@ export type ProviderDictionary = {
 };
 
 export type PreparedTransaction = {
-  predicate: FuelPredicate<InputValue[]>;
+  predicate: FuelPredicate<InputValue[], { [name: string]: unknown }>;
   request: TransactionRequest;
   transactionId: string;
   account: string;
@@ -61,3 +61,13 @@ export type PredicateCurrentState = {
   account?: string;
   accounts?: string[];
 };
+export interface PredicateVersionWithMetadata {
+  id: string;
+  generatedAt: number;
+  isActive: boolean;
+  isSelected: boolean;
+  isNewest: boolean;
+  balance?: string;
+  assetId?: string;
+  accountAddress?: string;
+}