+++ packages/react/src/ui/Connect/index.tsx
@@ -17,8 +17,12 @@ import { DialogContent } from './components/Core/DialogContent';
 import { DialogFuel } from './components/Core/DialogFuel';
 import { PredicateAddressDisclaimer } from './components/PredicateAddressDisclaimer/PredicateAddressDisclaimer';
 import { PredicateExternalDisclaimer } from './components/PredicateExternalDisclaimer/PredicateExternalDisclaimer';
+import { PredicateVersionDialog } from './components/PredicateVersion/PredicateVersionDialog';
+import { SignatureError } from './components/SignatureError/SignatureError';
 
 const ConnectRoutes = ({ route }: { route: Routes }) => {
+  const { theme } = useConnectUI();
+
   switch (route) {
     case Routes.List:
       return <Connectors />;
@@ -28,8 +32,12 @@ const ConnectRoutes = ({ route }: { route: Routes }) => {
       return <PredicateExternalDisclaimer />;
     case Routes.PredicateAddressDisclaimer:
       return <PredicateAddressDisclaimer />;
+    case Routes.PredicateVersionSelector:
+      return <PredicateVersionDialog theme={theme} />;
     case Routes.Connecting:
       return <Connecting />;
+    case Routes.SignatureError:
+      return <SignatureError theme={theme} />;
     default:
       return null;
   }